Refactor: voucher type을 확인하는 책임을 controller가 갖도록 변경
VoucherManager 가 IOManager를 알 수 없도록 레퍼런스 필드 삭제
