+++ src/main/java/prgms/vouchermanagementapp/io/CommandType.java
@@ -9,7 +9,7 @@ public enum CommandType {
 
     // TODO: UpperCase를 사용하여 리팩토링
     // 출력 역할 입장에서 생각해봐야한다.
-    EXIT("exit", "Type exit to exit the program.2"),
+    EXIT("exit", "Type exit to exit the program."),
     CREATE("create", "Type create to create a new voucher."),
     LIST("list", "Type list to list all vouchers.");
